Chore: rename *-util => *-utils (refs #10559)
