+++ lib/cli-engine.js
@@ -22,7 +22,7 @@ const fs = require("fs"),
     IgnoredPaths = require("./ignored-paths"),
     Config = require("./config"),
     LintResultCache = require("./util/lint-result-cache"),
-    globUtil = require("./util/glob-util"),
+    globUtil = require("./util/glob-utils"),
     validator = require("./config/config-validator"),
     hash = require("./util/hash"),
     ModuleResolver = require("./util/module-resolver"),
