+++ lib/config/config-initializer.js
@@ -16,9 +16,9 @@ const util = require("util"),
     autoconfig = require("./autoconfig.js"),
     ConfigFile = require("./config-file"),
     ConfigOps = require("./config-ops"),
-    getSourceCodeOfFiles = require("../util/source-code-util").getSourceCodeOfFiles,
+    getSourceCodeOfFiles = require("../util/source-code-utils").getSourceCodeOfFiles,
     ModuleResolver = require("../util/module-resolver"),
-    npmUtil = require("../util/npm-util"),
+    npmUtil = require("../util/npm-utils"),
     recConfig = require("../../conf/eslint-recommended"),
     log = require("../util/logging");
