+++ lib/ignored-paths.js
@@ -12,7 +12,7 @@
 const fs = require("fs"),
     path = require("path"),
     ignore = require("ignore"),
-    pathUtil = require("./util/path-util");
+    pathUtil = require("./util/path-utils");
 
 const debug = require("debug")("eslint:ignored-paths");
