+++ lib/util/naming.js
@@ -7,7 +7,7 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-const pathUtil = require("../util/path-util");
+const pathUtil = require("../util/path-utils");
 
 //------------------------------------------------------------------------------
 // Private
