+++ lib/util/npm-utils.js
