+++ lib/util/source-code-utils.js
@@ -10,10 +10,10 @@
 //------------------------------------------------------------------------------
 
 const CLIEngine = require("../cli-engine"),
-    globUtil = require("./glob-util"),
+    globUtil = require("./glob-utils"),
     baseDefaultOptions = require("../../conf/default-cli-options");
 
-const debug = require("debug")("eslint:source-code-util");
+const debug = require("debug")("eslint:source-code-utils");
 
 //------------------------------------------------------------------------------
 // Helpers
